refactor: blocks composable
