+++ apps/web/app/composables/useBlocks/types.ts
@@ -1,5 +1,4 @@
 import type { Block, GetBlocksResponse } from '@plentymarkets/shop-api';
-import type { FooterSwitchDefinition } from '~/components/blocks/Footer/types';
 
 export interface UseBlocksState {
   data: GetBlocksResponse;
@@ -27,7 +26,6 @@ export interface UseBlocks {
   discardChanges: () => void;
   setDefaultTemplate: (blocks: Block[]) => void;
   isSettling: Ref<boolean>;
-  FOOTER_SWITCH_DEFINITIONS: FooterSwitchDefinition[];
 }
 
 export type UseBlocksReturn = (localeOverride?: string) => UseBlocks;
